Move revenue growth function from utils to screen file
